vsf/requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
fake-useragent==0.1.11
ffmpy==0.3.0
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
h11==0.12.0
httmock==1.4.0
http-request-randomizer==1.3.2
httplib2==0.20.2
idna==3.3
numpy==1.22.0
oauthlib==3.1.1
outcome==1.1.0
pandas==1.3.5
protobuf==3.19.3
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pyOpenSSL==21.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
trio==0.19.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.8
webdriver-manager==3.5.2
wsproto==1.0.0
